+++ gazebo/gui/model/JointInspector.cc
@@ -108,6 +108,11 @@ JointInspector::JointInspector(JointMaker *_jointMaker, QWidget *_parent)
   this->configWidget->SetWidgetReadOnly("parent", true);
   this->configWidget->SetWidgetReadOnly("child", true);
 
+  // Get name widget
+  this->nameWidget = this->configWidget->ConfigChildWidgetByName("name");
+  if (!this->nameWidget)
+    gzerr << "Name widget not found" << std::endl;
+
   // Custom parent / child widgets
   // Parent
   std::vector<std::string> links;
@@ -187,8 +192,21 @@ JointInspector::JointInspector(JointMaker *_jointMaker, QWidget *_parent)
       SIGNAL(EnumValueChanged(const QString &, const QString &)), this,
       SLOT(OnEnumChanged(const QString &, const QString &)));
 
-  // Set initial joint type
-  this->OnJointTypeChanged(tr(msgs::Joint_Type_Name(jointMsg.type()).c_str()));
+  // Connect pose value changes, for joint pose
+  connect(this->configWidget, SIGNAL(PoseValueChanged(const QString &,
+      const ignition::math::Pose3d &)), this,
+      SLOT(OnPoseChanged(const QString &, const ignition::math::Pose3d &)));
+
+  // Connect vector value changes, for axes
+  connect(this->configWidget, SIGNAL(Vector3dValueChanged(const QString &,
+      const ignition::math::Vector3d &)), this,
+      SLOT(OnVector3dChanged(const QString &,
+      const ignition::math::Vector3d &)));
+
+  // Connect string value changes, for name
+  connect(this->configWidget, SIGNAL(StringValueChanged(const QString &,
+      const std::string &)), this, SLOT(OnStringChanged(const QString &,
+      const std::string &)));
 
   // Scroll area
   QScrollArea *scrollArea = new QScrollArea;
@@ -210,23 +228,21 @@ JointInspector::JointInspector(JointMaker *_jointMaker, QWidget *_parent)
   removeButton->setCheckable(false);
   connect(removeButton, SIGNAL(clicked()), this, SLOT(OnRemove()));
 
+  QPushButton *resetButton = new QPushButton(tr("Reset"));
+  connect(resetButton, SIGNAL(clicked()), this, SLOT(RestoreOriginalData()));
+
   QPushButton *cancelButton = new QPushButton(tr("Cancel"));
   connect(cancelButton, SIGNAL(clicked()), this, SLOT(OnCancel()));
 
-  this->applyButton = new QPushButton(tr("Apply"));
-  this->applyButton->setEnabled(true);
-  connect(this->applyButton, SIGNAL(clicked()), this, SLOT(OnApply()));
-
   this->okButton = new QPushButton(tr("OK"));
   this->okButton->setEnabled(true);
-  this->okButton->setDefault(true);
   connect(this->okButton, SIGNAL(clicked()), this, SLOT(OnOK()));
 
   QHBoxLayout *buttonsLayout = new QHBoxLayout;
   buttonsLayout->addWidget(removeButton);
   buttonsLayout->addStretch(5);
+  buttonsLayout->addWidget(resetButton);
   buttonsLayout->addWidget(cancelButton);
-  buttonsLayout->addWidget(this->applyButton);
   buttonsLayout->addWidget(this->okButton);
   buttonsLayout->setAlignment(Qt::AlignRight);
 
@@ -245,6 +261,14 @@ JointInspector::JointInspector(JointMaker *_jointMaker, QWidget *_parent)
       SLOT(OnLinkRemoved(std::string)));
   connect(this->jointMaker, SIGNAL(EmitLinkInserted(std::string)), this,
       SLOT(OnLinkInserted(std::string)));
+  connect(this, SIGNAL(rejected()), this, SLOT(RestoreOriginalData()));
+
+  // Initialize variables
+  this->validJointName = true;
+  this->validLinks = true;
+
+  // Set initial joint type
+  this->OnJointTypeChanged(tr(msgs::Joint_Type_Name(jointMsg.type()).c_str()));
 }
 
 /////////////////////////////////////////////////
@@ -282,6 +306,11 @@ msgs::Joint *JointInspector::GetData() const
 
   // Get updated message from widget
   msgs::Joint *msg = dynamic_cast<msgs::Joint *>(this->configWidget->GetMsg());
+  if (!msg)
+  {
+    gzerr << "It wasn't possible to get the joint message" << std::endl;
+    return NULL;
+  }
 
   // Use parent / child from our custom widget
   msg->set_parent(currentParent);
@@ -298,6 +327,48 @@ void JointInspector::OnEnumChanged(const QString &_name,
     this->OnJointTypeChanged(_value);
   else if (_name == "parentCombo" || _name == "childCombo")
     this->OnLinksChanged(_value);
+}
+
+/////////////////////////////////////////////////
+void JointInspector::OnPoseChanged(const QString &/*_name*/,
+    const ignition::math::Pose3d &/*_pose*/)
+{
+  if (this->CheckValid())
+    emit Applied();
+}
+
+/////////////////////////////////////////////////
+void JointInspector::OnVector3dChanged(const QString &/*_name*/,
+    const ignition::math::Vector3d &/*_vec*/)
+{
+  if (this->CheckValid())
+    emit Applied();
+}
+
+/////////////////////////////////////////////////
+void JointInspector::OnStringChanged(const QString &_name,
+    const std::string &_str)
+{
+  // Handle joint name
+  if (_name != "name")
+    return;
+
+  /// \todo Also check if name overlaps with other joints
+  this->validJointName = !_str.empty();
+
+  // Warning if joint name is invalid
+  if (!this->validJointName)
+  {
+    this->nameWidget->setStyleSheet(ConfigWidget::StyleSheet("warning"));
+  }
+  else
+  {
+    this->nameWidget->setStyleSheet(ConfigWidget::StyleSheet("normal"));
+  }
+
+  // Only apply if all fields are valid
+  if (this->CheckValid())
+    emit Applied();
 }
 
 /////////////////////////////////////////////////
@@ -350,6 +421,9 @@ void JointInspector::OnJointTypeChanged(const QString &_value)
           (matAmbient[2] * 255) << "); }";
 
   this->parentIcon->setStyleSheet(QString::fromStdString(sheet.str()));
+
+  if (this->CheckValid())
+    emit Applied();
 }
 
 /////////////////////////////////////////////////
@@ -371,8 +445,11 @@ void JointInspector::OnLinksChanged(const QString &/*_linkName*/)
     this->parentLinkWidget->setStyleSheet(ConfigWidget::StyleSheet("normal"));
     this->childLinkWidget->setStyleSheet(ConfigWidget::StyleSheet("normal"));
   }
-  this->applyButton->setEnabled(currentParent != currentChild);
-  this->okButton->setEnabled(currentParent != currentChild);
+  this->validLinks = currentParent != currentChild;
+
+  // Only apply if all fields are valid
+  if (this->CheckValid())
+    emit Applied();
 }
 
 /////////////////////////////////////////////////
@@ -384,8 +461,10 @@ void JointInspector::OnSwap()
   std::string currentChild =
       this->configWidget->GetEnumWidgetValue("childCombo");
 
-  // Choose new values
+  // Choose new values. We only need signals to be emitted once.
+  this->configWidget->blockSignals(true);
   this->configWidget->SetEnumWidgetValue("parentCombo", currentChild);
+  this->configWidget->blockSignals(false);
   this->configWidget->SetEnumWidgetValue("childCombo", currentParent);
 }
 
@@ -444,8 +523,13 @@ void JointInspector::Open()
   // Reset states
   this->parentLinkWidget->setStyleSheet(ConfigWidget::StyleSheet("normal"));
   this->childLinkWidget->setStyleSheet(ConfigWidget::StyleSheet("normal"));
-  this->applyButton->setEnabled(true);
+  this->nameWidget->setStyleSheet(ConfigWidget::StyleSheet("normal"));
   this->okButton->setEnabled(true);
+
+  // Keep original data in case user cancels
+  auto msg = this->GetData();
+  if (msg)
+    this->originalDataMsg.CopyFrom(*msg);
 
   this->move(QCursor::pos());
   this->show();
@@ -466,26 +550,75 @@ void JointInspector::OnRemove()
 }
 
 /////////////////////////////////////////////////
-void JointInspector::OnCancel()
+void JointInspector::RestoreOriginalData()
 {
-  this->close();
+  msgs::JointPtr jointPtr;
+  jointPtr.reset(new msgs::Joint);
+  jointPtr->CopyFrom(this->originalDataMsg);
+
+  // Update default widgets
+  this->configWidget->blockSignals(true);
+  this->Update(jointPtr);
+
+  // Update joint type and parent icon
+  this->OnJointTypeChanged(tr(msgs::Joint_Type_Name(jointPtr->type()).c_str()));
+
+  // Update custom widgets
+  this->configWidget->SetEnumWidgetValue("parentCombo",
+      this->configWidget->GetStringWidgetValue("parent"));
+  this->configWidget->SetEnumWidgetValue("childCombo",
+      this->configWidget->GetStringWidgetValue("child"));
+  this->configWidget->blockSignals(false);
+
+  if (this->CheckValid())
+    emit Applied();
 }
 
 /////////////////////////////////////////////////
-void JointInspector::OnApply()
+void JointInspector::OnCancel()
 {
-  emit Applied();
+  this->RestoreOriginalData();
+
+  this->close();
 }
 
 /////////////////////////////////////////////////
 void JointInspector::OnOK()
 {
-  emit Applied();
-  this->accept();
+  if (this->CheckValid())
+  {
+    emit Applied();
+    this->accept();
+  }
+  else
+  {
+    gzerr << "It shouldn't be possible to press Ok with invalid inputs." <<
+        std::endl;
+  }
 }
 
 /////////////////////////////////////////////////
 void JointInspector::enterEvent(QEvent */*_event*/)
 {
   QApplication::setOverrideCursor(Qt::ArrowCursor);
+}
+
+/////////////////////////////////////////////////
+bool JointInspector::CheckValid()
+{
+  bool valid = this->validJointName && this->validLinks;
+
+  if (this->okButton)
+    this->okButton->setEnabled(valid);
+
+  return valid;
+}
+
+/////////////////////////////////////////////////
+void JointInspector::keyPressEvent(QKeyEvent *_event)
+{
+  if (_event->key() == Qt::Key_Enter)
+    _event->accept();
+  else
+    QDialog::keyPressEvent(_event);
 }